Wither and Hunger
Instead of damage being applied when exposed to the sky now players get wither applied. Also made it so that when the player is above 65 y in the overworld they get hunger applied. This will hopefully prevent players from staying on the surface for too long even if not exposed to the sky.